fix compiler warning in CCBReader

+++ cocos2d-ui/CCBReader/CCBReader.m
@@ -172,7 +172,7 @@ static inline unsigned int readVariableLengthIntFromArray(const uint8_t* buffer,
     
 done:
     *value = result;
-    return ptr - buffer;
+    return (unsigned int) (ptr - buffer);
 }
